trackWidgets.js build out.
