+++ index.html
@@ -324,38 +324,6 @@
         </div>
     </div>
 
-    <!-- Generic Track Select Modal -->
-    <div id="igv-app-generic-track-select-modal" class="modal fade">
-
-        <div class="modal-dialog modal-lg">
-
-            <div class="modal-content">
-
-                <div class="modal-header">
-                    <div id="igv-app-generic-track-select-modal-label" class="modal-title">
-                    </div>
-
-                    <button type="button" class="close" data-dismiss="modal">
-                        <span aria-hidden="true">&times;</span>
-                    </button>
-
-                </div>
-
-                <div class="modal-body">
-
-                    <div class="form-group">
-                        <select class="form-control">
-                        </select>
-                    </div>
-
-                </div>
-
-            </div>
-
-        </div>
-
-    </div>
-
     <!-- Genome from URL -->
     <div id="igv-app-genome-from-url-modal" class="modal fade">
 
@@ -365,36 +333,6 @@
 
                 <div class="modal-header">
                     <div id="igv-app-genome-from-url-modal-label" class="modal-title">Genome URL</div>
-
-                    <button type="button" class="close" data-dismiss="modal">
-                        <span aria-hidden="true">&times;</span>
-                    </button>
-
-                </div>
-
-                <div class="modal-body">
-                </div>
-
-                <div class="modal-footer">
-                    <button type="button" class="btn btn-sm btn-outline-secondary" data-dismiss="modal">Cancel</button>
-                    <button type="button" class="btn btn-sm btn-secondary">OK</button>
-                </div>
-
-            </div>
-
-        </div>
-
-    </div>
-
-    <!-- Track from URL -->
-    <div id="igv-app-track-from-url-modal" class="modal fade">
-
-        <div class="modal-dialog modal-lg">
-
-            <div class="modal-content">
-
-                <div class="modal-header">
-                    <div id="igv-app-track-from-url-modal-label" class="modal-title">Track URL</div>
 
                     <button type="button" class="close" data-dismiss="modal">
                         <span aria-hidden="true">&times;</span>
